Extract `test_module_dependencies_complete` into its own file
